test: re-tier deterministic poisoning/whitebox-pim tests into the fast suite
BadNets (pure trigger-stamping, no training) and WhiteBoxPIM's
attack-before-prepare precondition check (raises before any computation)
were marked @pytest.mark.integration despite running in milliseconds on
synthetic data, so they only ran on the slow path instead of every push.
